backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.88.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
ecdsa==0.19.2
fastapi==0.135.3
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
pyasn1==0.6.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
